Move ErrorMessage to core-common
