+++ app/src/main/kotlin/com/hedvig/app/feature/embark/quotecart/CreateQuoteCartUseCase.kt
@@ -6,10 +6,10 @@ import com.hedvig.android.apollo.graphql.CreateOnboardingQuoteCartMutation
 import com.hedvig.android.apollo.graphql.type.Market
 import com.hedvig.android.apollo.safeExecute
 import com.hedvig.android.apollo.toEither
+import com.hedvig.android.core.common.ErrorMessage
 import com.hedvig.android.language.LanguageService
 import com.hedvig.android.market.MarketManager
 import com.hedvig.app.feature.offer.model.QuoteCartId
-import com.hedvig.app.util.ErrorMessage
 
 class CreateQuoteCartUseCase(
   private val apolloClient: ApolloClient,
