+++ app/src/main/kotlin/com/hedvig/app/feature/home/ui/changeaddress/GetAddressChangeStoryIdUseCase.kt
@@ -6,11 +6,11 @@ import com.apollographql.apollo3.ApolloClient
 import com.hedvig.android.apollo.graphql.ActiveContractBundlesQuery
 import com.hedvig.android.apollo.safeExecute
 import com.hedvig.android.apollo.toEither
+import com.hedvig.android.core.common.ErrorMessage
 import com.hedvig.android.hanalytics.featureflags.FeatureManager
 import com.hedvig.android.hanalytics.featureflags.flags.Feature
 import com.hedvig.app.feature.embark.QUOTE_CART_EMBARK_STORE_ID_KEY
 import com.hedvig.app.feature.embark.quotecart.CreateQuoteCartUseCase
-import com.hedvig.app.util.ErrorMessage
 
 class GetAddressChangeStoryIdUseCase(
   private val apolloClient: ApolloClient,
