+++ app/src/main/kotlin/com/hedvig/app/feature/insurance/data/GetContractsUseCase.kt
@@ -6,8 +6,8 @@ import com.apollographql.apollo3.ApolloClient
 import com.hedvig.android.apollo.graphql.InsuranceQuery
 import com.hedvig.android.apollo.safeExecute
 import com.hedvig.android.apollo.toEither
+import com.hedvig.android.core.common.ErrorMessage
 import com.hedvig.android.language.LanguageService
-import com.hedvig.app.util.ErrorMessage
 
 class GetContractsUseCase(
   private val apolloClient: ApolloClient,
