+++ app/src/main/kotlin/com/hedvig/app/feature/offer/ui/changestartdate/QuoteCartEditStartDateUseCase.kt
@@ -7,9 +7,9 @@ import com.apollographql.apollo3.ApolloClient
 import com.hedvig.android.apollo.graphql.QuoteCartEditQuoteMutation
 import com.hedvig.android.apollo.safeExecute
 import com.hedvig.android.apollo.toEither
+import com.hedvig.android.core.common.ErrorMessage
 import com.hedvig.android.language.LanguageService
 import com.hedvig.app.feature.offer.model.QuoteCartId
-import com.hedvig.app.util.ErrorMessage
 import kotlinx.serialization.json.buildJsonObject
 import kotlinx.serialization.json.put
 import kotlinx.serialization.json.putJsonObject
