+++ app/src/main/kotlin/com/hedvig/app/feature/offer/usecase/EditCampaignUseCase.kt
@@ -8,9 +8,9 @@ import com.hedvig.android.apollo.graphql.QuoteCartAddCampaignMutation
 import com.hedvig.android.apollo.graphql.QuoteCartRemoveCampaignMutation
 import com.hedvig.android.apollo.safeExecute
 import com.hedvig.android.apollo.toEither
+import com.hedvig.android.core.common.ErrorMessage
 import com.hedvig.app.feature.offer.OfferRepository
 import com.hedvig.app.feature.offer.model.QuoteCartId
-import com.hedvig.app.util.ErrorMessage
 
 @JvmInline
 value class CampaignCode(val code: String)
