+++ core-common/src/main/kotlin/com/hedvig/android/core/common/ErrorMessage.kt
@@ -1,4 +1,4 @@
-package com.hedvig.app.util
+package com.hedvig.android.core.common
 
 @JvmInline
 value class ErrorMessage(val message: String? = null)
